duui-mm-embedder-gradle/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.0.1
av==15.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
decorator==5.2.1
deprecation==2.1.0
dkpro-cassis==0.10.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.6
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
lxml==5.3.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numba==0.62.0
numpy==2.3.3
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
psutil==7.0.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.38.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
toposort==1.10
torch==2.8.0
torchaudio==2.8.0
torchcodec==0.7.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
